feat: Refactor files add a new comp Chat.jsx
